Add geometry module inside the service module, to centralize geometric calculations
